Support assignment to parameters in the dart2dart backend.
Assignments to parameters update the assigned variables during CPS
construction.  This has the effect of propagating the assigned value
to its uses and the original assignment cannot necessarily be
recovered in the backend.

This enables multiple uses of the same let-bound primitive in the CPS
IR, which requires handling by introducing an assignment in the Dart
output.  The final code is not yet as compact as it could be.  For
instance:

hest(x, y) {
  x = fisk(x);
  hest(x, x);
}

is translated to:

hest(x,y){var v0;v0=fisk(x);hest(v0,v0);}

R=floitsch@google.com, karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org//233153002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@34964 260f80e4-7a28-3924-810f-c04153c831b5