Allowing Angular sample to run on dev branch with Angular 0.9.10
This overrides Angular's analyzer version which is used for command-line tools only to match the version used by DI 0.36.0, so that the published version of Angular will work on the Dart dev branch.

The dependency override only works now because the analyzer dependency is only for the command-line utilities which are being phased out in favor of transformers (appearing in next Angular version).

BUG=
R=nweiz@google.com

Review URL: https://codereview.chromium.org//235933014

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@35056 260f80e4-7a28-3924-810f-c04153c831b5