Add default selector exclution flag to test.dart
This can be used like this:
tools/test.py -mrelease --exclude-suite=co19,pkg
to run the default selector tests but not run pkg and co19 tests.

Why am I doing this:
We want to keep broad testing on developer machines, but want to allow testing resources to have a better and more logic distribution on the bots, e.g., not run pkg tests on the vm bots.

R=whesse@google.com

Review URL: https://codereview.chromium.org//256703003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@35424 260f80e4-7a28-3924-810f-c04153c831b5