Optimize conditional branches that have same true/false targets.
Branches that have the same true/false target, or where both
target blocks reach the same common join block via empty blocks
can be replaced by a goto to the common join block.

In code like this

var a = unknown();
var b = null;
if (a == null || b == null) {
    ...
}

it eliminates the test (a == null) if b is known to be null. Until now,
the compiler could only eliminate the test for b, if a was known.

R=vegorov@google.com

Review URL: https://codereview.chromium.org//261823005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@35669 260f80e4-7a28-3924-810f-c04153c831b5