Handle constructor initializers in constant evaluation.
Since one library may contain a constant instance creation expression that refers to a class in another library, we may not have the ASTs for the constructor initializer expressions at the time we evaluate the constant.  So we clone the initializer AST's into the element model to keep them available.

This cloning must happen after evaluating any constants referred to by the initializer AST's, but before evaluating any constants that make use of the constructor.  So we add the appropriate dependencies to ConstantValueComputer's dependency graph, and do the cloning in ConstantValueComputer.computeValueFor().

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//272293004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@36043 260f80e4-7a28-3924-810f-c04153c831b5