Fix code invalidation of functions with references to deferred libraries
When a code object with a reference to a deferred library prefix is
invalidated, it must be removed from the function object, and also
patched so that direct calls to it (from optimized code for example)
will trap and redirect to the newly compiled code.

Removing an overly eager assertion.

This fixes bug 18909.

R=fschneider@google.com, srdjan@google.com

Review URL: https://codereview.chromium.org//296853002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@36436 260f80e4-7a28-3924-810f-c04153c831b5