Fix dartbug.com/18725 by avoiding to rewrite into the console at process launch time (writes apparently appear on the stdin of the launched process).
This is not necessarily the fix you want to go with, but I thought I'd send it your
way because it fixes the (pretty bad) issue that puts extra stuff on the stdin for
processes. I also like that the output from the process is cleanly separated from
the editor UI bits (showing the cmdline). Ideally, we can also make the Observatory
print (from the VM) go away so the output in the console is completely under the
app's control.

R=devoncarew@google.com, keertip@google.com
BUG=dartbug.com/18725

Review URL: https://codereview.chromium.org//299323004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@36643 260f80e4-7a28-3924-810f-c04153c831b5