Simplify computeMergedExecutableElement() in light of r36579.
Prior to r36579, the spec allowed for an impossible condition in which
a synthetic element for an inherited method could not be created.  The
impossible condition has now been dropped from the spec, so change the
analyzer accordingly.

No functional change since the condition could never arise.

R=jwren@google.com

Review URL: https://codereview.chromium.org//292313013

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@36691 260f80e4-7a28-3924-810f-c04153c831b5