Avoid infinite loops in analysis when a file can't be read.
Previously, if AnalysisContextImpl.CycleBuilder's ensureImports or
ensureExports method encountered a file that was in an error state
(because it couldn't be read from the filesystem), it would still try
to resolve it, leading to an exception being thrown and no analysis
progress being made.  As a result, analysis would get stuck in an
infinite loop trying to analyze the unreadable file over and over
again.

(This situation could sometimes arise even in the file was only
unreadable for a short period of time, e.g. because of a race
condition between analysis and a source control operation).

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//372423002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@38074 260f80e4-7a28-3924-810f-c04153c831b5