Do not notify debugger on stack overflow
Don’t call back to the debugger on out of memory and stack overflow
exceptions. The debugger can’t do anything meaningful on the VM
when the stack is full.

This should eliminate crashes we see in Dartium devtools.

This means we can no longer intercept stack overflows with the debugger.
In a later change, I plan to add uncomment the code that notifies
the debugger of the exception after the stack in unwound. Before this
notification gets uncommented, we need to make sure that debuggers
can handle notifications with an empty stack.

R=asiva@google.com

Review URL: https://codereview.chromium.org//376053003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@38086 260f80e4-7a28-3924-810f-c04153c831b5