Redundant phi elimination in new backend
This commit adds a redundant phi elimination pass over the IR in the new
backend. Whenever an argument to all invokes of a continuation is the
same (except for feedback), the argument can be eliminated both from
calls and the the continuation signature, replacing all usages within
the continuation body.

BUG=
R=floitsch@google.com

Review URL: https://codereview.chromium.org//398823007

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@38492 260f80e4-7a28-3924-810f-c04153c831b5