Bugfix in html_dart2js: Detect window objects in a way compatible with cross domain iframes
The "setTimeout" member in window object can only be accessed if the window
object is from the same origin. Cross origin windows will throw a security error
exception.

This is fixed by checking for "postMessage" in the window object (which
can be accessed even in cross-origin windows).

BUG=http://dartbug.com/20166
R=blois@google.com

Review URL: https://codereview.chromium.org//406143004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@38517 260f80e4-7a28-3924-810f-c04153c831b5