Start running the analysis server integration tests by changing the name to end with _test.dart
Note that these will not run on the vm bots, only on the pub/pkg bots on the right.

I would rather that we keep this in the normal setup. If we are seeing to long cycle times we should shard these machines instead of special casing.

Sorry if you wasted time doing the new specialized test suite.

If we want to, we can skip this on specific platforms in the status files(qua comments in mail thread), but I don't think we should

I ran this in a loop on my linux box for a few hours, no flakes.

R=whesse@google.com

Review URL: https://codereview.chromium.org//426923002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@38661 260f80e4-7a28-3924-810f-c04153c831b5