dart2dart: Swap the order of phi elimination and variable allocation.
Variable allocation should be run last after all transformations on
the CPS IR.  It allocates IR variables to Tree variables based on an
approximation of IR variable liveness.  This approximation can be
invalidated by transformation of the IR.

This covered by test/language/liveness_test.dart if a redundant phi is
introduced on the for loop back edge, but impossible (?) to produce
with the current IR builder.

BUG=
R=jgruber@google.com

Review URL: https://codereview.chromium.org//444613002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@38896 260f80e4-7a28-3924-810f-c04153c831b5