Browser testing: Run android content shell tests on hosts with multiple devices connected.
The adb_reverse_forwarder.py script in [dartium]/build/android/ needs the --all-devices option added. We will also try to upstream this flag to chromium.
This change makes the test host reinstall the ContentShell apk on the device each time the device is restarted due to timeout or periodic reset, which takes about 10 seconds, but this is simpler and more like the existing behavior with android chrome.

R=ricow@google.com

Review URL: https://codereview.chromium.org//467643002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@39142 260f80e4-7a28-3924-810f-c04153c831b5