dart2dart: Add support for continue in loops.
For `while` loops, continue is just a jump to the top of the loop.
For `for` loops the situation is more complicated: continue is a
forward jump to the update expression.

The scoping of the continue continuation would ideally be at the
common ancestor of all its invocations, but that is hard to compute in
a single pass and without parent pointers.  Instead, the continue
continuation is inserted at a point that is safe (in scope for the
entire body) at the cost of passing all locals as parameters.

BUG=
R=floitsch@google.com

Review URL: https://codereview.chromium.org//480513002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@39348 260f80e4-7a28-3924-810f-c04153c831b5