Fix Issue 20342: type propagation for Object methods on dynamic exps
Tests are included.

This solution seems hackish, in that it's not generic in the API of
[Object]: it could break if new APIs were added to the [Object] class
-- e.g. new operators defined on [Object] would be broken. However, we
don't expect [Object]'s API to change, so this is probably a safe
solution.

An outstanding question: should we include tooltip documentation for
[Object] methods/properties resolved on dynamic expressions? Right now
we do not. A reason to continue not showing docs here is that
subclasses can override the docs when they override the method --
e.g. the [hashCode] docs for [int] are different from those for
[Object] -- so showing docs for [Object] methods on dynamic
expressions could be misleading.  Moreover, it probably doesn't
matter: people should already know what [==], [hashCode], and
[toString] do, and [runtimeType] should be easy to guess ...

R=brianwilkerson@google.com, jwren@google.com
BUG=https://code.google.com/p/dart/issues/detail?id=20342

Review URL: https://codereview.chromium.org//511513002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@39683 260f80e4-7a28-3924-810f-c04153c831b5