VM: Use current isolate as immediate operand in per-isolate stub code.
There is no need to load it from the context, if the stub
is per-isolate. This makes those stubs not rely on having a context in CTX,
and avoids loading the isolate pointer from memory.

This CL is part of the effort to free up CTX for general
allocation. Stubs that don't depend on a context in CTX
won't require storing the current context before calling
them.

R=srdjan@google.com

Review URL: https://codereview.chromium.org//581703003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@40428 260f80e4-7a28-3924-810f-c04153c831b5