first cut fuzz test for analysis server
This very rough first cut drives the analysis server snapshot so that it...

- analyzes the specified source tree
- locates and 'opens' a file in that tree
- modifies the content of that file in memory without saving to disk
- requests code completions

I'm not happy with the readability of the ServerTest # test() method because of all the Future references.
I'm going to experiment with async / await to see if that can make it more readable.

There are several libraries that I pulled out of the analysis server code base and cleaned up to use
with an analysis server client. In the long run, these analysis server client libraries need to be cleaned up
so that clients can import them rather than having to do what I've done. These libraries include:

- example/server_test/bin/byte_stream_channel.dart
- example/server_test/bin/channel.dart
- example/server_test/bin/generated_protocol.dart
- example/server_test/bin/json.dart
- example/server_test/bin/protocol.dart

BUG=
R=lukechurch@google.com

Review URL: https://codereview.chromium.org//584963002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@40556 260f80e4-7a28-3924-810f-c04153c831b5