Implement strict and non-strict union types behind flag [strictUnionTypes].
This only covers existing union types code. In particular, there is
still no support for fields on union types (treated as dynamic
currently, so no hints ever).

It remains to implement strict and non-strict versions of the
completion UI. That's in progress and will go up in another CL because
this CL was getting too big.

R=brianwilkerson@google.com, jwren@google.com
BUG=

Review URL: https://codereview.chromium.org//599293002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@40679 260f80e4-7a28-3924-810f-c04153c831b5